+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.7.0] - 2021-02-09
 ### Added
 - instance health check plugin
+- vinyl metrics
 
 ## [0.6.1] - 2021-01-20
 ### Fixed